Created all entities.

+++ DKK/DKK/Schemas/Person/Tables/Competitor.sql
@@ -4,7 +4,7 @@
 	,PersonId INT NOT NULL CONSTRAINT FK_Person_Person_PersonId FOREIGN KEY REFERENCES Person.Person (PersonId)
 	,DateOfBirth DATE NULL
 	,Age TINYINT NOT NULL
-	,[Weight] SMALLINT NOT NULL
+	,[Weight] DECIMAL(5,2) NOT NULL
 	,RankId INT NOT NULL CONSTRAINT FK_Person_Rank_RankId FOREIGN KEY REFERENCES [Event].[Rank] (RankId)
 	,DojoId INT NOT NULL CONSTRAINT FK_Facility_Dojo_DojoId FOREIGN KEY REFERENCES Facility.Dojo (DojoId)
 	,ParentId INT NULL CONSTRAINT FK_Person_Person_ParentId FOREIGN KEY REFERENCES Person.Person (PersonId)
